Port over old search tests.
